+++ types/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 1.0.0-rc.5 - YYYY-MM-DD
+
+### Added
+
+- `NativeTokens::{contains, get}`;
+
 ## 1.0.0-rc.4 - 2022-12-20
 
 ### Added